Add moduletest for OsmModule
